Add stream conformance tests for Deflate/ZLib/GZip/BrotliStream
